+++ src/Training.AirBnb.Clone.Frontend/src/modules/locations/components/ListingCard.vue
