+++ src/Training.AirBnb.Clone.Frontend/src/modules/locations/components/ListingsCategoryTab.vue
