+++ src/Training.AirBnb.Clone.Frontend/src/modules/profile/components/DarkModeToggler.vue
@@ -1,0 +1,23 @@
+<template>
+
+    <div :class="modeValue ? 'bg-logoPrimary' : 'bg-gray-300'" @click="toggle"
+         class="group w-10 h-6 flex primary-transition-toggle theme-bg-accent rounded-full p-1 theme-text-primary">
+        <div :class="modeValue ? 'translate-x-4' : 'translate-x-0'"
+             class="w-4 h-4 primary-transition  theme-bg-toggle rounded-full theme-shadow"></div>
+    </div>
+
+</template>
+
+<script setup lang="ts">
+import {ref} from "vue";
+import {AppThemeService} from "@/infrastructure/services/AppThemeService";
+
+const appThemeService = new AppThemeService();
+const modeValue = ref<boolean>(appThemeService.isDarkMode())
+
+const toggle = () => {
+    modeValue.value = !modeValue.value;
+    appThemeService.toggleDarkMode();
+}
+
+</script>
